Continue work on porting components to React
Ported SelectionDialog and fixed a lot of issues with actions and the
  reducer.

Things will probably have to be rewritten to be at least half decent,
but at least there is some progress being made.

Almost sure that only the Sidebar fixing and the Schedule itself are
missing. Almost done with porting, and later on would be able to get
into unspaghetting.