Setup development proxy
This allows using the tts production api in case the developer doesn't want to run it locally.
This was a problem because apparently it does not have CORS setup
correctly.

However, this is not an issue since the idea is to migrate away from it.
Additionally, this will also be useful for working with sigarra's api,
since it also does not provide decent CORS setup (even the mobile api!!).