Fixes the handling of courses that have dots in their ID
This caused the confirm button to be disabled.
A potential future fix may be related to using a different separator or
key to access nested attributes.

Potential issues may arise from the use of `Vue.set` with an array key
(second argument) since the doc does not support it.

However, since objects in javascript can handle being accessed by key
using variables of type Array, it should work fine (the value is
serialized to `a,b,c` like array.join(',')).